[Zeppelin-1229] Add cache buster for zeppelin-web build
### What is this PR for?
In order to bust the cache when there is changes to the code of zeppelin-web after build, we are adding hashes at the end of the imported .js and .css files we use.
Every time there is changes in one of those file, the hash will change and the browser will use new file instead of cached one.

### What type of PR is it?
Improvement

### What is the Jira issue?
https://issues.apache.org/jira/browse/ZEPPELIN-1229

### How should this be tested?
* Build zeppelin-web once and write down the HASH in `dist/scripts/scripts.HASH.js`
* After making a change inside a .js file of the application and building again, the hash in `dist/scripts/scripts.HASH.js` should be changed.

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Damien CORNEAU <corneadoug@gmail.com>

Closes #1225 from corneadoug/ZEPPELIN-1229 and squashes the following commits:

ebad090 [Damien CORNEAU] Add cache buster for .js and .css file in zeppelin-web production build